requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.3
asgiref==3.6.0
attrs==23.2.0
beautifulsoup4==4.12.2
black==23.3.0
blinker==1.7.0
boto3==1.28.40
botocore==1.31.40
bs4==0.0.1
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
dataclasses-json==0.6.4
debugpy==1.6.7
defusedxml==0.7.1
diff-match-patch==20230430
distro==1.9.0
Django==4.1.7
django-apscheduler==0.6.2
django-cors-headers==3.14.0
django-crontab==0.7.1
django-filter==23.1
django-import-export==3.3.1
django-rest-swagger==2.2.0
django-ses==3.5.0
django-smtp-ssl==1.0
django-suit==0.2.28
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
faiss-cpu==1.7.4
Flask==3.0.2
Flask-Cors==4.0.0
frozenlist==1.4.1
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.4
import-export==0.3.1
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.9
langchain-community==0.0.21
langchain-core==0.1.26
langchain-openai==0.0.6
langsmith==0.1.5
Markdown==3.4.3
MarkupPy==1.14
MarkupSafe==2.1.3
marshmallow==3.20.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.24.4
odfpy==1.4.1
openai==1.12.0
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==23.2
pandas==2.2.1
pathspec==0.11.1
platformdirs==3.2.0
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==1.10.14
pydantic_core==2.16.2
PyJWT==1.7.1
PyMySQL==1.1.0
pypdf==4.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
python-slugify==8.0.4
pytz==2023.2
PyYAML==6.0.1
regex==2023.12.25
requests==2.28.2
s3transfer==0.6.2
schedule==1.2.1
sendgrid==6.10.0
simplejson==3.19.2
six==1.16.0
slugify==0.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.27
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
stripe==7.3.0
sysrsync==1.1.0
tablib==3.5.0
tenacity==8.2.3
text-unidecode==1.3
tiktoken==0.6.0
toml==0.10.2
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.15
Werkzeug==3.0.1
whitenoise==6.4.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.4
